requirements.txt
# Market Dashboard Requirements

# Google Sheets API
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Gemini API
google-generativeai>=0.3.0

# X (Twitter) API
tweepy>=4.14.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP Requests
requests>=2.31.0
curl_cffi>=0.5.0

# Environment Variables
python-dotenv>=1.0.0
